refine paddle_inference_helper.h
