+++ paddle/fluid/inference/api/helper.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "paddle/fluid/inference/api/helper.h"
+#include "paddle/fluid/inference/api/paddle_inference_helper.h"
 
 namespace paddle {
 namespace inference {
